Update sidebar to use shared component
